Created platform_utils.py
